Fixed bug in wellformed_prereqs and added 1 verification
The bug was due to our updated version of prereqs. Now since the classes are an OR relation within equivalent courses it is possible for there to be a legal cycle. For example, the prereqs for 1570 are 1010 or 0220. The prereqs for 1010 are 1570, 0220, or 1550. Although 1570 and 1010 are in each other's prereqs, there is a path that is not a cycle, making this legal.